feat: add time in force label
