+++ libs/deal-ticket/src/components/deal-ticket/time-in-force-selector.tsx
@@ -2,40 +2,13 @@ import { useEffect, useState } from 'react';
 import { FormGroup, Select } from '@vegaprotocol/ui-toolkit';
 import { Schema } from '@vegaprotocol/types';
 import { t } from '@vegaprotocol/react-helpers';
+import { timeInForceLabel } from '@vegaprotocol/orders';
 
 interface TimeInForceSelectorProps {
   value: Schema.OrderTimeInForce;
   orderType: Schema.OrderType;
   onSelect: (tif: Schema.OrderTimeInForce) => void;
 }
-
-// More detail in https://docs.vega.xyz/docs/mainnet/graphql/enums/order-time-in-force
-export const timeInForceLabel = (tif: string) => {
-  switch (tif) {
-    case Schema.OrderTimeInForce.TIME_IN_FORCE_GTC:
-      return t(`Good 'til Cancelled (GTC)`);
-    case Schema.OrderTimeInForce.TIME_IN_FORCE_IOC:
-      return t('Immediate or Cancel (IOC)');
-    case Schema.OrderTimeInForce.TIME_IN_FORCE_FOK:
-      return t('Fill or Kill (FOK)');
-    case Schema.OrderTimeInForce.TIME_IN_FORCE_GTT:
-      return t(`Good 'til Time (GTT)`);
-    case Schema.OrderTimeInForce.TIME_IN_FORCE_GFN:
-      return t('Good for Normal (GFN)');
-    case Schema.OrderTimeInForce.TIME_IN_FORCE_GFA:
-      return t('Good for Auction (GFA)');
-    default:
-      return t(tif);
-  }
-};
-
-type PossibleOrderKeys = Exclude<
-  Schema.OrderType,
-  Schema.OrderType.TYPE_NETWORK
->;
-type PrevSelectedValue = {
-  [key in PossibleOrderKeys]: Schema.OrderTimeInForce;
-};
 
 export const TimeInForceSelector = ({
   value,
