+++ libs/orders/src/lib/index.ts
@@ -1,2 +1,3 @@
 export * from './components';
 export * from './order-hooks';
+export * from './uitls';
